+++ test/component/composer/cases/case08.js
@@ -1,0 +1,5 @@
+export const title       = '8) Handle events';
+export const description = 'You can capture with <code>addEventListener</code> the regular events for Graphery Components and HTMLElements.';
+export const events      = true;
+
+export default '<g-composer svg-src="/test/component/composer/assets/check.svg" style="width: 50px;"></g-composer>';