+++ test/component/template/cases/case09.js
@@ -1,4 +1,0 @@
-export const title       = '9) Hidden attribute';
-export const description = 'You can capture hidden the component with the hidden property (and attribute).';
-
-export default `<g-template svg-src="/test/component/template/assets/check.svg" $[selector: g-template; attribute: hidden; type: boolean; value: true] style="width: 50px;"></g-template>`;