+++ test/component/template/cases/case18.js
@@ -1,8 +1,0 @@
-export const title       = '18) render with data-src';
-export const description = 'data-src and data-type attributs';
-
-export default `<g-template        
-  svg-src="/test/component/template/assets/circles.svg"
-  data-src="/test/component/template/assets/circles.csv">
-</g-template>
-`;